Splitting up a test file
